refactor: migrate balance components and clean up documentation
Rename balance/components_old to balance/components and streamline documentation:
- Rename frontend/src/features/balance/components_old/ → components/
- Simplify BalanceDisplay.tsx comments by removing verbose examples and remarks
- Keep essential documentation while removing redundant explanations

This completes the balance feature component migration as part of Phase 2 directory structure cleanup.
