+++ frontend/src/features/balance/components/__stories__/BalanceDisplay.stories.tsx
